refactor: 반복되는 컴포넌트 리팩토링
